requirements.txt
torch
transformers
datasets
scikit-learn
numpy
pandas
hydra-core
omegaconf
wandb
python-dotenv
tqdm
pytest
umap-learn
matplotlib
accelerate